fix some lint issues

+++ app/routes/notes/$noteId.tsx
@@ -2,7 +2,6 @@ import type { ActionFunctionArgs, LoaderFunctionArgs } from "react-router";
 import {
   redirect,
   Form,
-  isRouteErrorResponse,
   useLoaderData,
   useRouteError,
 } from "react-router";
@@ -11,21 +10,24 @@ import invariant from "tiny-invariant";
 import { Box } from "~/components/Box/Box";
 import { Button } from "~/components/Button/Button";
 import { deleteNote, getNote } from "~/models/note.server";
-import { requireUserId } from "~/session.server";
+import { getUserId, loginRedirect } from "~/session.server";
+import { HttpError } from "~/utils/httpError.server";
 
 export const loader = async ({ params, request, url }: LoaderFunctionArgs) => {
-  const userId = await requireUserId(request, url);
+  const userId = await getUserId(request);
+  if (!userId) return loginRedirect(url);
   invariant(params.noteId, "noteId not found");
 
   const note = await getNote({ id: params.noteId, userId });
   if (!note) {
-    throw new Response("Not Found", { status: 404 });
+    throw new HttpError("Not Found", 404);
   }
   return { note };
 };
 
 export const action = async ({ params, request, url }: ActionFunctionArgs) => {
-  const userId = await requireUserId(request, url);
+  const userId = await getUserId(request);
+  if (!userId) return loginRedirect(url);
   invariant(params.noteId, "noteId not found");
 
   await deleteNote({ id: params.noteId, userId });
@@ -51,17 +53,13 @@ export default function NoteDetailsPage() {
 export function ErrorBoundary() {
   const error = useRouteError();
 
+  if (error instanceof HttpError && error.status === 404) {
+    return <Box>Note not found</Box>;
+  }
+
   if (error instanceof Error) {
     return <Box>An unexpected error occurred: {error.message}</Box>;
   }
 
-  if (!isRouteErrorResponse(error)) {
-    return <h1>Unknown Error</h1>;
-  }
-
-  if (error.status === 404) {
-    return <Box>Note not found</Box>;
-  }
-
-  return <Box>An unexpected error occurred: {error.statusText}</Box>;
+  return <h1>Unknown Error</h1>;
 }
